Add badge_type_id as an attribute of an Athlete
The Strava API is returning an integer field called "badge_type_id"
when I request my athlete details, which is generating a warning
because this library is not aware of the attribute.

There is no API documentation at all for this attribute, so I'm not
able to add any details about where it appears or what it means.