Make to_dict() include attributes of parent classes
Before this change `to_dict()` would only return values that were defined on the
local class.

For example, `Bike.to_dict()` would return a dict with a single `frame_type` key
since it's the only `Attribute` defined on the class. The rest of the
attributes are defined on the `Gear`, `IdentifiableEntity`, etc.

This commit makes `to_dict()` also look for `Attributes` defined on parent
classes and include them in the output as well.